refactor: automated code improvements for maintainability
- extract MemoryGraph widget from settings_page.dart to reduce file size and apply Single Responsibility
- create _handle_db_errors async context manager in ProductivityService to consolidate duplicate try/catch blocks and apply DRY principles
