+++ backend/app/domain/productivity/service.py
@@ -3,6 +3,8 @@
 from __future__ import annotations
 
 import logging
+from collections.abc import AsyncGenerator
+from contextlib import asynccontextmanager
 from uuid import UUID
 
 from fastapi import HTTPException
@@ -23,6 +25,21 @@ from app.domain.productivity.models import (
 logger = logging.getLogger(__name__)
 
 
+@asynccontextmanager
+async def _handle_db_errors(operation_name: str) -> AsyncGenerator[None, None]:
+    """Handle database errors consistently across productivity services."""
+    try:
+        yield
+    except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
+        logger.exception(f"Failed to {operation_name}")
+        raise HTTPException(
+            status_code=500, detail=f"Database error occurred while {operation_name}"
+        ) from e
+    except Exception as e:
+        logger.exception(f"Unexpected error in {operation_name}")
+        raise HTTPException(status_code=500, detail=f"Failed to {operation_name}") from e
+
+
 class ProductivityService:
     """Service for managing productivity features."""
 
@@ -33,40 +50,24 @@ class ProductivityService:
     @staticmethod
     async def create_task(user_id: UUID, task_data: TaskCreate) -> Task:
         """Create a new task for the user."""
-        try:
+        async with _handle_db_errors("create task"):
             return await Task.create(
                 user_id=user_id,
                 title=task_data.title,
                 description=task_data.description,
                 is_completed=task_data.is_completed,
             )
-        except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
-            logger.exception("Failed to create task")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while creating task"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in create_task")
-            raise HTTPException(status_code=500, detail="Failed to create task") from e
 
     @staticmethod
     async def list_tasks(user_id: UUID, limit: int = 50, offset: int = 0) -> list[Task]:
         """List tasks for the user with pagination."""
-        try:
+        async with _handle_db_errors("list tasks"):
             return (
                 await Task.filter(user_id=user_id)
                 .order_by("-created_at")
                 .limit(limit)
                 .offset(offset)
             )
-        except (IntegrityError, OperationalError, DBConnectionError) as e:
-            logger.exception("Failed to list tasks")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while listing tasks"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in list_tasks")
-            raise HTTPException(status_code=500, detail="Failed to list tasks") from e
 
     @staticmethod
     async def get_task(user_id: UUID, task_id: UUID) -> Task:
@@ -95,34 +96,18 @@ class ProductivityService:
         if not valid_keys:
             return task
 
-        try:
+        async with _handle_db_errors("update task"):
             for field, value in valid_keys.items():
                 setattr(task, field, value)
             await task.save(update_fields=list(valid_keys.keys()))
             return task
-        except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
-            logger.exception("Failed to update task")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while updating task"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in update_task")
-            raise HTTPException(status_code=500, detail="Failed to update task") from e
 
     @staticmethod
     async def delete_task(user_id: UUID, task_id: UUID) -> None:
         """Delete a specific task."""
         task = await ProductivityService.get_task(user_id, task_id)
-        try:
+        async with _handle_db_errors("delete task"):
             await task.delete()
-        except (IntegrityError, OperationalError, DBConnectionError) as e:
-            logger.exception("Failed to delete task")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while deleting task"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in delete_task")
-            raise HTTPException(status_code=500, detail="Failed to delete task") from e
 
     # ---------------------------------------------------------------------------
     # Notes
@@ -131,7 +116,7 @@ class ProductivityService:
     @staticmethod
     async def create_note(user_id: UUID, note_data: NoteCreate) -> Note:
         """Create a new note for the user."""
-        try:
+        async with _handle_db_errors("create note"):
             return await Note.create(
                 user_id=user_id,
                 agent_id=note_data.agent_id,
@@ -142,33 +127,16 @@ class ProductivityService:
                 is_pinned=note_data.is_pinned,
             )
 
-        except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
-            logger.exception("Failed to create note")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while creating note"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in create_note")
-            raise HTTPException(status_code=500, detail="Failed to create note") from e
-
     @staticmethod
     async def list_notes(user_id: UUID, limit: int = 50, offset: int = 0) -> list[Note]:
         """List notes for the user, pinned first, then by creation date."""
-        try:
+        async with _handle_db_errors("list notes"):
             return (
                 await Note.filter(user_id=user_id)
                 .order_by("-is_pinned", "-created_at")
                 .limit(limit)
                 .offset(offset)
             )
-        except (IntegrityError, OperationalError, DBConnectionError) as e:
-            logger.exception("Failed to list notes")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while listing notes"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in list_notes")
-            raise HTTPException(status_code=500, detail="Failed to list notes") from e
 
     @staticmethod
     async def get_note(user_id: UUID, note_id: UUID) -> Note:
@@ -197,34 +165,18 @@ class ProductivityService:
         if not valid_keys:
             return note
 
-        try:
+        async with _handle_db_errors("update note"):
             for field, value in valid_keys.items():
                 setattr(note, field, value)
             await note.save(update_fields=list(valid_keys.keys()))
             return note
-        except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
-            logger.exception("Failed to update note")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while updating note"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in update_note")
-            raise HTTPException(status_code=500, detail="Failed to update note") from e
 
     @staticmethod
     async def delete_note(user_id: UUID, note_id: UUID) -> None:
         """Delete a specific note."""
         note = await ProductivityService.get_note(user_id, note_id)
-        try:
+        async with _handle_db_errors("delete note"):
             await note.delete()
-        except (IntegrityError, OperationalError, DBConnectionError) as e:
-            logger.exception("Failed to delete note")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while deleting note"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in delete_note")
-            raise HTTPException(status_code=500, detail="Failed to delete note") from e
 
     # ---------------------------------------------------------------------------
     # Calendar Events
@@ -237,7 +189,7 @@ class ProductivityService:
             logger.warning("Attempted to create event with end_time before start_time")
             raise HTTPException(status_code=400, detail="end_time must be greater than start_time")
 
-        try:
+        async with _handle_db_errors("create calendar event"):
             return await CalendarEvent.create(
                 user_id=user_id,
                 agent_id=event_data.agent_id,
@@ -250,33 +202,17 @@ class ProductivityService:
                 is_all_day=event_data.is_all_day,
                 location=event_data.location,
             )
-        except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
-            logger.exception("Failed to create calendar event")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while creating calendar event"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in create_event")
-            raise HTTPException(status_code=500, detail="Failed to create calendar event") from e
 
     @staticmethod
     async def list_events(user_id: UUID, limit: int = 50, offset: int = 0) -> list[CalendarEvent]:
         """List calendar events for the user."""
-        try:
+        async with _handle_db_errors("list calendar events"):
             return (
                 await CalendarEvent.filter(user_id=user_id)
                 .order_by("start_time")
                 .limit(limit)
                 .offset(offset)
             )
-        except (IntegrityError, OperationalError, DBConnectionError) as e:
-            logger.exception("Failed to list calendar events")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while listing calendar events"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in list_events")
-            raise HTTPException(status_code=500, detail="Failed to list calendar events") from e
 
     @staticmethod
     async def get_event(user_id: UUID, event_id: UUID) -> CalendarEvent:
@@ -314,31 +250,15 @@ class ProductivityService:
             logger.warning("Attempted to update event with end_time before start_time")
             raise HTTPException(status_code=400, detail="end_time must be greater than start_time")
 
-        try:
+        async with _handle_db_errors("update calendar event"):
             for field, value in valid_keys.items():
                 setattr(event, field, value)
             await event.save(update_fields=list(valid_keys.keys()))
             return event
-        except (IntegrityError, OperationalError, DBConnectionError, ValueError) as e:
-            logger.exception("Failed to update calendar event")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while updating calendar event"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in update_event")
-            raise HTTPException(status_code=500, detail="Failed to update calendar event") from e
 
     @staticmethod
     async def delete_event(user_id: UUID, event_id: UUID) -> None:
         """Delete a specific calendar event."""
         event = await ProductivityService.get_event(user_id, event_id)
-        try:
+        async with _handle_db_errors("delete calendar event"):
             await event.delete()
-        except (IntegrityError, OperationalError, DBConnectionError) as e:
-            logger.exception("Failed to delete calendar event")
-            raise HTTPException(
-                status_code=500, detail="Database error occurred while deleting calendar event"
-            ) from e
-        except Exception as e:
-            logger.exception("Unexpected error in delete_event")
-            raise HTTPException(status_code=500, detail="Failed to delete calendar event") from e
