+++ frontend/lib/features/settings/pages/settings_page.dart
@@ -1,5 +1,4 @@
 import 'dart:developer' as developer;
-import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -10,6 +9,7 @@ import 'package:miru/core/services/passkey_service.dart';
 import 'package:miru/core/services/supabase_service.dart';
 import 'package:miru/core/api/api_service.dart';
 import 'package:miru/core/models/memory.dart';
+import 'package:miru/features/settings/widgets/memory_graph.dart';
 
 class SettingsPage extends StatefulWidget {
   /// Called when the user clears chat history from settings.
@@ -360,7 +360,10 @@ class _SettingsPageState extends State<SettingsPage> {
             else ...[
               _buildMemoryViewToggle(),
               if (_showMemoryGraph)
-                _buildMemoryGraph()
+                MemoryGraphWidget(
+                  memories: _memories,
+                  memoryEdges: _memoryEdges,
+                )
               else
                 ..._memories.map(_buildMemoryListItem),
             ],
@@ -605,50 +608,6 @@ class _SettingsPageState extends State<SettingsPage> {
     );
   }
 
-  Widget _buildMemoryGraph() {
-    final colors = context.colors;
-
-    if (_memories.length == 1) {
-      return Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: AppSpacing.lg,
-          vertical: AppSpacing.sm,
-        ),
-        child: Text(
-          'Only one memory available. More connections appear as Miru learns.',
-          style: AppTypography.bodySmall.copyWith(color: colors.onSurfaceMuted),
-        ),
-      );
-    }
-
-    final nodes = _memories.map(_GraphNode.fromMemory).toList();
-
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(
-        AppSpacing.lg,
-        AppSpacing.xs,
-        AppSpacing.lg,
-        AppSpacing.sm,
-      ),
-      child: Container(
-        height: 280,
-        decoration: BoxDecoration(
-          color: colors.surface,
-          borderRadius: BorderRadius.circular(AppSpacing.md),
-          border: Border.all(color: colors.border.withValues(alpha: 0.3)),
-        ),
-        child: CustomPaint(
-          painter: _MemoryGraphPainter(
-            colors: colors,
-            nodes: nodes,
-            edges: _memoryEdges,
-          ),
-          child: const SizedBox.expand(),
-        ),
-      ),
-    );
-  }
-
   Widget _buildSectionHeader(String title) {
     return Padding(
       padding: const EdgeInsets.fromLTRB(
@@ -703,104 +662,5 @@ class _SettingsPageState extends State<SettingsPage> {
         vertical: AppSpacing.xs,
       ),
     );
-  }
-}
-
-class _GraphNode {
-  final String id;
-  final String content;
-
-  const _GraphNode({required this.id, required this.content});
-
-  factory _GraphNode.fromMemory(Memory memory) {
-    return _GraphNode(id: memory.id, content: memory.content);
-  }
-}
-
-class _MemoryGraphPainter extends CustomPainter {
-  final AppThemeColors colors;
-  final List<_GraphNode> nodes;
-  final List<MemoryEdge> edges;
-
-  const _MemoryGraphPainter({
-    required this.colors,
-    required this.nodes,
-    required this.edges,
-  });
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    if (nodes.isEmpty) {
-      return;
-    }
-
-    final nodePositions = <String, Offset>{};
-    final center = Offset(size.width / 2, size.height / 2);
-    final radius = math.min(size.width, size.height) * 0.33;
-
-    for (var i = 0; i < nodes.length; i++) {
-      final angle = (2 * math.pi * i) / nodes.length;
-      nodePositions[nodes[i].id] = Offset(
-        center.dx + radius * math.cos(angle),
-        center.dy + radius * math.sin(angle),
-      );
-    }
-
-    final edgePaint = Paint()
-      ..color = colors.primary.withValues(alpha: 0.32)
-      ..strokeWidth = 1.6
-      ..style = PaintingStyle.stroke;
-
-    for (final edge in edges) {
-      final sourceId = edge.source;
-      final targetId = edge.target;
-
-      final sourcePoint = nodePositions[sourceId];
-      final targetPoint = nodePositions[targetId];
-      if (sourcePoint == null || targetPoint == null) {
-        continue;
-      }
-
-      canvas.drawLine(sourcePoint, targetPoint, edgePaint);
-    }
-
-    for (final node in nodes) {
-      final point = nodePositions[node.id];
-      if (point == null) {
-        continue;
-      }
-
-      final nodePaint = Paint()..color = colors.primary;
-      canvas.drawCircle(point, 8, nodePaint);
-
-      final label = _shortLabel(node.content);
-      final textPainter = TextPainter(
-        text: TextSpan(
-          text: label,
-          style: AppTypography.bodySmall.copyWith(color: colors.onSurface),
-        ),
-        textDirection: TextDirection.ltr,
-        maxLines: 1,
-        ellipsis: '...',
-      )..layout(maxWidth: 110);
-
-      final dx = point.dx - textPainter.width / 2;
-      final dy = point.dy + 12;
-
-      textPainter.paint(canvas, Offset(dx, dy));
-    }
-  }
-
-  String _shortLabel(String content) {
-    final trimmed = content.trim();
-    if (trimmed.length <= 18) {
-      return trimmed;
-    }
-    return '${trimmed.substring(0, 18)}...';
-  }
-
-  @override
-  bool shouldRepaint(covariant _MemoryGraphPainter oldDelegate) {
-    return oldDelegate.nodes != nodes || oldDelegate.edges != edges;
   }
 }
